refactor: move `DocumentAttributes` to its own file
